+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.0
- - Added a column in the generated html report that shows the mutation match pattern
+ - Added optional "id" attribute to rules in the xml configuration file
+ - Added a column in the generated html report that shows the mutation pattern and id
 
 ## 1.3.4
  - Improved mutation rules that switch function arguments